coerce bools to strings
because we have several layers of yaml, templating, and interpolation going on with manifests, runtime config, bosh vars files, credhub, etc, sometimes our bools are ending up as strings. If we coerce whatever our variables are into strings, we know they'll always be strings and we can compare them as such.